+++ FrontEnd/src/App.css
@@ -50,24 +50,6 @@
   font-size: 20px;
 }
 
-/* .navbar-cap-1 {
-  margin: 0;
-  font-size: 40px;
-  margin-top: 30px;
-  width: fit-content;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.navbar-cap-2 {
-  margin: 0;
-  font-size: 20px;
-  margin-top: 30px;
-  width: fit-content;
-  margin-left: auto;
-  margin-right: auto;
-} */
-
 .hidden {
   display: none;
 }
@@ -122,4 +104,59 @@
 
 .rotate {
   transform: rotate(360deg);
+}
+
+/* Dropdown Menu styles for buttons */
+.dropdown-menubutton, .dropdown-settingsbutton {
+  position: absolute;
+  top: 85%; /* Position it below the settings button */
+  background-color: #D9D9D9;
+  border-radius: 10px;
+  box-shadow: 0 8px 16px rgba(0, 0, 0, 0.1);
+  z-index: 20; /* Make sure it's above other elements */
+  width: 150px;
+  overflow: hidden; /* Ensures the border-radius clips the content */
+}
+
+.dropdown-menubutton{
+  left: 0;
+}
+
+.dropdown-settingsbutton{
+  right: 0;
+}
+
+.nav .dropdown-menubutton {
+  margin-left: 2rem; /* Margin */
+}
+
+.nav .dropdown-settingsbutton {
+  margin-right: 2rem; /* Margin */
+}
+
+/* Style for each item in the dropdown menu */
+.dropdown-menubutton a, .dropdown-settingsbutton a {
+  display: block;
+  background-color: #808080;
+  color: white;
+  border-radius: 15px;
+  top: 10px;
+  width: calc(100% - 20px);
+  padding: 5px;
+  margin: 5px 10px;
+  text-decoration: none;
+  transition: #154734 0.2s;
+  text-align: center;
+  font-weight: bold;
+  box-sizing: border-box;
+}
+
+/* Remove border from the last item */
+.dropdown-menubutton a:last-child, .dropdown-settingsbutton a:last-child {
+  border-bottom: none;
+}
+
+/* Hover effect for each item */
+.dropdown-menubutton a:hover, .dropdown-settingsbutton a:hover {
+  background-color: #154734;
 }
